include separate npm scripts for running adr tests
