+++ src/app/static/package.json
@@ -8,8 +8,10 @@
     "js": "webpack",
     "build": "gulp copy-templates && gulp styles && webpack",
     "test": "jest -c jest.config.js",
-    "integration-test": "jest -c jest.integration.config.js",
-    "integration-test-docker": "jest -c jest.integration.docker.config.js",
+    "integration-test": "jest --testPathIgnorePatterns adr -c jest.integration.config.js",
+    "adr-integration-test": "jest --testPathPattern adr -c jest.integration.config.js",
+    "integration-test-docker": "jest --testPathIgnorePatterns adr -c jest.integration.docker.config.js",
+    "adr-integration-test-docker": "jest --testPathPattern adr -c jest.integration.docker.config.js",
     "lint": "npx eslint src --ext .js,.jsx,.ts,.tsx,vue"
   },
   "author": "",
@@ -54,7 +56,7 @@
     "html-webpack-plugin": "^3.2.0",
     "i18next": "^19.0.1",
     "ify-loader": "^1.1.0",
-    "jest": "^25.0.0",
+    "jest": "^26.6.3",
     "jest-canvas-mock": "^2.1.2",
     "jest-teamcity-reporter": "^0.9.0",
     "jquery": "^3.5.1",
